academy api 생성 GET /api/academy/check-dupl POST /api/academy/new POST /api/academy/verification
