+++ src/main/java/psam/portfolio/sunder/english/global/advice/exception/WebControllerAdvice.java
@@ -15,7 +15,7 @@ import org.springframework.web.multipart.MultipartException;
 import psam.portfolio.sunder.english.global.api.ApiException;
 import psam.portfolio.sunder.english.global.api.ApiResponse;
 import psam.portfolio.sunder.english.global.api.ApiStatus;
-import psam.portfolio.sunder.english.web.user.model.User;
+import psam.portfolio.sunder.english.web.user.model.entity.User;
 
 import static org.springframework.http.HttpStatus.*;
