+++ src/main/java/psam/portfolio/sunder/english/global/security/UserDetailsServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import psam.portfolio.sunder.english.web.user.model.User;
+import psam.portfolio.sunder.english.web.user.model.entity.User;
 import psam.portfolio.sunder.english.web.user.exception.NoSuchUserException;
 import psam.portfolio.sunder.english.web.user.repository.UserQueryRepository;
