+++ src/main/java/psam/portfolio/sunder/english/web/teacher/model/request/AcademyPATCH.java
@@ -1,4 +1,4 @@
 package psam.portfolio.sunder.english.web.teacher.model.request;
 
-public class TeacherUpdateInfo {
+public class AcademyPATCH {
 }
