+++ src/main/java/psam/portfolio/sunder/english/web/user/controller/UserController.java
@@ -16,8 +16,6 @@ import java.util.Map;
 public class UserController {
 
     // TODO: 2024-01-26 Guide 문서 만들기.
-    // TODO: 2024-01-26 build.gradle 에서 test config 다듬기
-    // TODO: 2024-01-26 User 중복 검사 API 문서 - email, phone case 만들기
 
     private final UserQueryService userQueryService;
