+++ src/main/java/psam/portfolio/sunder/english/web/user/repository/UserCommandRepository.java
@@ -1,7 +1,7 @@
 package psam.portfolio.sunder.english.web.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import psam.portfolio.sunder.english.web.user.model.User;
+import psam.portfolio.sunder.english.web.user.model.entity.User;
 
 import java.util.UUID;
