+++ src/main/java/psam/portfolio/sunder/english/web/user/repository/UserRoleCommandRepository.java
@@ -1,7 +1,7 @@
 package psam.portfolio.sunder.english.web.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import psam.portfolio.sunder.english.web.user.model.UserRole;
+import psam.portfolio.sunder.english.web.user.model.entity.UserRole;
 
 public interface UserRoleCommandRepository extends JpaRepository<UserRole, Long> {
 }
