+++ src/test/java/psam/portfolio/sunder/english/service/AcademyCommandServiceTest.java
@@ -2,17 +2,15 @@ package psam.portfolio.sunder.english.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import psam.portfolio.sunder.english.SunderApplicationTests;
 import psam.portfolio.sunder.english.infrastructure.password.PasswordUtils;
 import psam.portfolio.sunder.english.web.teacher.enumeration.AcademyStatus;
 import psam.portfolio.sunder.english.web.teacher.exception.DuplicateAcademyException;
-import psam.portfolio.sunder.english.web.teacher.exception.OneParamToCheckAcademyDuplException;
 import psam.portfolio.sunder.english.web.teacher.model.entity.Academy;
 import psam.portfolio.sunder.english.web.teacher.model.entity.Teacher;
-import psam.portfolio.sunder.english.web.teacher.model.request.AcademyRegistration;
-import psam.portfolio.sunder.english.web.teacher.model.request.DirectorRegistration;
+import psam.portfolio.sunder.english.web.teacher.model.request.AcademyDirectorPOST.DirectorPOST;
+import psam.portfolio.sunder.english.web.teacher.model.request.AcademyDirectorPOST.AcademyPOST;
 import psam.portfolio.sunder.english.web.teacher.repository.AcademyQueryRepository;
 import psam.portfolio.sunder.english.web.teacher.repository.TeacherQueryRepository;
 import psam.portfolio.sunder.english.web.teacher.service.AcademyCommandService;
@@ -21,9 +19,10 @@ import psam.portfolio.sunder.english.web.user.exception.DuplicateUserException;
 
 import java.util.UUID;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.BDDMockito.*;
+import static org.mockito.BDDMockito.anyString;
+import static org.mockito.BDDMockito.given;
 
 @SuppressWarnings("ConstantValue")
 public class AcademyCommandServiceTest extends SunderApplicationTests {
@@ -40,101 +39,6 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
     @Autowired
     PasswordUtils passwordUtils;
 
-    /*
-     * 1. 학원의 정보 중복 검사를 할 때는 하나의 데이터만 전달해야 한다.
-     * 2. 학원의 정보 중복 검사를 할 때는 아이디, 이메일, 연락처 중 하나는 반드시 전달해야 한다.
-     * 3. 학원 이름의 중복 검사를 수행할 수 있다.
-     * 4. 학원의 연락처 중복 검사를 수행할 수 있다.
-     * 5. 학원의 이메일 중복 검사를 수행할 수 있다.
-     *
-     * 6. 학원을 등록할 때도 학원 정보의 중복 검사를 수행한다.
-     * 7. 학원을 등록하면서 선생의 정보도 중복 검사를 수행한다.
-     * 8. PENDING 상태의 학원은 중복 검사에서 제외된다.
-     * 9. 등록된 학원장의 비밀번호는 암호화되어 있다.
-     * 10. 등록된 학원과 학원장은 이메일 인증이 필요하다.
-     * 12. 학원 uuid 로 학원과 학원장을 인증할 수 있다.
-     * 12. 학원은 최초 1회만 인증할 수 있다.
-     */
-
-    @DisplayName("학원의 정보 중복 검사를 할 때는 하나의 데이터만 전달해야 한다.")
-    @Test
-    void oneParamToCheckDuplException(){
-        // given
-        String name = "name";
-        String phone = "";
-        String email = "email";
-
-        // when
-        // then
-        assertThatThrownBy(() -> runWithRefresh(() -> sut.checkDuplication(name, phone, email)))
-                .isInstanceOf(OneParamToCheckAcademyDuplException.class);
-    }
-
-    @DisplayName("학원의 정보 중복 검사를 할 때는 아이디, 이메일, 연락처 중 하나는 반드시 전달해야 한다.")
-    @Test
-    void noParamToCheckDuplException(){
-        // given
-        String name = "";
-        String phone = "";
-        String email = "";
-
-        // when
-        // then
-        assertThatThrownBy(() -> runWithRefresh(() -> sut.checkDuplication(name, phone, email)))
-                .isInstanceOf(OneParamToCheckAcademyDuplException.class);
-    }
-
-    @DisplayName("학원 이름의 중복 검사를 수행할 수 있다.")
-    @Test
-    void checkNameDupl(){
-        // given
-        Academy registerAcademy = registerAcademy(AcademyStatus.VERIFIED);
-
-        String name = registerAcademy.getName();
-        String phone = null;
-        String email = null;
-
-        // when
-        boolean isOk = runWithRefresh(() -> sut.checkDuplication(name, phone, email));
-
-        // then
-        assertThat(isOk).isFalse();
-    }
-
-    @DisplayName("학원의 연락처 중복 검사를 수행할 수 있다.")
-    @Test
-    void checkPhoneDupl(){
-        // given
-        Academy registerAcademy = registerAcademy(AcademyStatus.VERIFIED);
-
-        String name = null;
-        String phone = registerAcademy.getPhone();
-        String email = null;
-
-        // when
-        boolean isOk = runWithRefresh(() -> sut.checkDuplication(name, phone, email));
-
-        // then
-        assertThat(isOk).isFalse();
-    }
-
-    @DisplayName("학원의 이메일 중복 검사를 수행할 수 있다.")
-    @Test
-    void checkEmailDupl(){
-        // given
-        Academy registerAcademy = registerAcademy(AcademyStatus.VERIFIED);
-
-        String name = null;
-        String phone = null;
-        String email = registerAcademy.getEmail();
-
-        // when
-        boolean isOk = runWithRefresh(() -> sut.checkDuplication(name, phone, email));
-
-        // then
-        assertThat(isOk).isFalse();
-    }
-
     @DisplayName("학원을 등록할 때도 학원 정보의 중복 검사를 수행한다.")
     @Test
     void checkDuplWhenRegisterAcademy(){
@@ -145,7 +49,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
         // given
         Academy duplicateAcademy = registerAcademy(AcademyStatus.VERIFIED);
 
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(duplicateAcademy.getName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -155,7 +59,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(uic.getUniqueId())
                 .loginPw("loginPw")
                 .name("name")
@@ -168,7 +72,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
 
         // when
         // then
-        assertThatThrownBy(() -> runWithRefresh(() -> sut.registerDirectorWithAcademy(academyRegistration, directorRegistration)))
+        assertThatThrownBy(() -> runWithRefresh(() -> sut.registerDirectorWithAcademy(academyPOST, directorPOST)))
                 .isInstanceOf(DuplicateAcademyException.class);
     }
 
@@ -183,7 +87,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
         Academy registerAcademy = registerAcademy(AcademyStatus.VERIFIED);
         Teacher duplicateTeacher = registerTeacher(UserStatus.ACTIVE, registerAcademy);
 
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(uic.getUniqueAcademyName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -193,7 +97,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(duplicateTeacher.getLoginId())
                 .loginPw("loginPw")
                 .name("name")
@@ -206,25 +110,8 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
 
         // when
         // then
-        assertThatThrownBy(() -> runWithRefresh(() -> sut.registerDirectorWithAcademy(academyRegistration, directorRegistration)))
+        assertThatThrownBy(() -> runWithRefresh(() -> sut.registerDirectorWithAcademy(academyPOST, directorPOST)))
                 .isInstanceOf(DuplicateUserException.class);
-    }
-
-    @DisplayName("PENDING 상태의 학원은 중복 검사에서 제외된다.")
-    @Test
-    void ifPendingOk(){
-        // given
-        Academy academy = registerAcademy(AcademyStatus.PENDING);
-
-        String name = academy.getName();
-        String email = null;
-        String phone = null;
-
-        // when
-        boolean isOk = runWithRefresh(() -> sut.checkDuplication(name, phone, email));
-
-        // then
-        assertThat(isOk).isTrue();
     }
 
     @DisplayName("등록된 학원장의 비밀번호는 암호화되어 있다.")
@@ -235,7 +122,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .willReturn(true);
 
         // given
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(uic.getUniqueAcademyName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -245,7 +132,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(uic.getUniqueId())
                 .loginPw("loginPw")
                 .name("name")
@@ -257,7 +144,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .build();
 
         // when
-        String teacherUuid = runWithRefresh(() -> sut.registerDirectorWithAcademy(academyRegistration, directorRegistration));
+        String teacherUuid = runWithRefresh(() -> sut.registerDirectorWithAcademy(academyPOST, directorPOST));
 
         // then
         Teacher getTeacher = teacherQueryRepository.getById(UUID.fromString(teacherUuid));
@@ -272,7 +159,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .willReturn(true);
 
         // given
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(uic.getUniqueAcademyName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -282,7 +169,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(uic.getUniqueId())
                 .loginPw("loginPw")
                 .name("name")
@@ -294,7 +181,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .build();
 
         // when
-        String teacherUuid = runWithRefresh(() -> sut.registerDirectorWithAcademy(academyRegistration, directorRegistration));
+        String teacherUuid = runWithRefresh(() -> sut.registerDirectorWithAcademy(academyPOST, directorPOST));
 
         // then
         Teacher getTeacher = teacherQueryRepository.getById(UUID.fromString(teacherUuid));
@@ -312,7 +199,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .willReturn(true);
 
         // given
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(uic.getUniqueAcademyName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -322,7 +209,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(uic.getUniqueId())
                 .loginPw("loginPw")
                 .name("name")
@@ -333,7 +220,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .postalCode("00000")
                 .build();
 
-        String teacherUuid = sut.registerDirectorWithAcademy(academyRegistration, directorRegistration);
+        String teacherUuid = sut.registerDirectorWithAcademy(academyPOST, directorPOST);
         Teacher getTeacher = teacherQueryRepository.getById(UUID.fromString(teacherUuid));
         UUID academyUuid = getTeacher.getAcademy().getUuid();
 
@@ -358,7 +245,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .willReturn(true);
 
         // given
-        AcademyRegistration academyRegistration = AcademyRegistration.builder()
+        AcademyPOST academyPOST = AcademyPOST.builder()
                 .name(uic.getUniqueAcademyName())
                 .phone(uic.getUniquePhoneNumber())
                 .email(uic.getUniqueEmail())
@@ -368,7 +255,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .openToPublic(true)
                 .build();
 
-        DirectorRegistration directorRegistration = DirectorRegistration.builder()
+        DirectorPOST directorPOST = DirectorPOST.builder()
                 .loginId(uic.getUniqueId())
                 .loginPw("loginPw")
                 .name("name")
@@ -379,7 +266,7 @@ public class AcademyCommandServiceTest extends SunderApplicationTests {
                 .postalCode("00000")
                 .build();
 
-        String teacherUuid = sut.registerDirectorWithAcademy(academyRegistration, directorRegistration);
+        String teacherUuid = sut.registerDirectorWithAcademy(academyPOST, directorPOST);
         Teacher getTeacher = teacherQueryRepository.getById(UUID.fromString(teacherUuid));
         UUID academyUuid = getTeacher.getAcademy().getUuid();
         runWithRefresh(() -> sut.verifyAcademy(academyUuid));
